requirements.txt
sounddevice>=0.4.6
pynput>=1.7.6
vosk>=0.3.45
numpy>=1.21.0
pystray>=0.19.5
Pillow>=9.0.0
